+++ src/platforms/rcore_desktop_rgfw.c
@@ -1024,15 +1024,15 @@ Image GetClipboardImage(void)
 #if SUPPORT_CLIPBOARD_IMAGE && SUPPORT_MODULE_RTEXTURES
 #if defined(_WIN32)
 
-    unsigned long long int dataSize = 0; // moved into _WIN32 scope until other platforms gain support
-    void *fileData = NULL; // moved into _WIN32 scope until other platforms gain support
+    unsigned int dataSize = 0;
+    void *fileData = NULL;
 
     int width = 0;
     int height = 0;
-    fileData  = (void *)Win32GetClipboardImageData(&width, &height, &dataSize);
+    fileData = (void *)Win32GetClipboardImageData(&width, &height, &dataSize);
 
     if (fileData == NULL) TRACELOG(LOG_WARNING, "Clipboard image: Couldn't get clipboard data");
-    else image = LoadImageFromMemory(".bmp", (const unsigned char *)fileData, dataSize);
+    else image = LoadImageFromMemory(".bmp", (const unsigned char *)fileData, (int)dataSize);
 
 #elif defined(__linux__) && defined(DRGFW_X11)
 
@@ -1082,7 +1082,7 @@ Image GetClipboardImage(void)
     XCloseDisplay(dpy);
 #else
     TRACELOG(LOG_WARNING, "Clipboard image: PLATFORM_DESKTOP_RGFW doesn't implement GetClipboardImage() for this OS");
-#endif // defined(_WIN32)
+#endif // _WIN32
 #else
     TRACELOG(LOG_WARNING, "Clipboard image: SUPPORT_CLIPBOARD_IMAGE requires SUPPORT_MODULE_RTEXTURES to work properly");
 #endif // SUPPORT_CLIPBOARD_IMAGE
